Change import from beta version to stable version
