+++ src/components/ProjectCard/index.jsx
@@ -6,6 +6,7 @@ import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
+import { Link } from 'react-router-dom';
 import Markdown from 'react-markdown';
 
 @withStyles(theme => ({
